refactor: moved pypi installation into its own module
1. Moved files
2. Clippy fixed some small issues
